add Bio.IO (and clean up)
